Fix Android onMomentumScrollEnd when interrupted
Interrupting a momentum scroll by dragging will incorrectly fire two
`onMomentumScrollEnd` events, both of which happen too early. This can
be fixed by stopping the fling when the user starts dragging again and
checking whether the fling is still going on in the runnable that
eventually detects when a fling has ended.

Fixes #18568.